(Refactor) Refactored ross to the new lib structure.
