+++ icevision/models/ross/efficientdet/backbones/__init__.py
@@ -1,0 +1,1 @@
+from icevision.models.ross.efficientdet.backbones.backbones import *
