--- icevision/models/ross/efficientdet/fastai/__init__.py
@@ -1,2 +1,0 @@
-from icevision.models.ross.efficientdet.fastai.callbacks import *
-from icevision.models.ross.efficientdet.fastai.learner import *
