--- icevision/models/ross/efficientdet/lightning/__init__.py
@@ -1,1 +1,0 @@
-from icevision.models.ross.efficientdet.lightning.model_adapter import *
